build tao bridge with bazel
