+++ scripts/ci/build_and_test.sh
@@ -30,16 +30,16 @@ BLADE_DISC_DIR=tao/python/blade_disc_tf
   && rm -rf tao/build \
   && cd tf_community && bazel clean --expunge)
 
-python ${ENTRY} ${VENV_PATH} -s configure --bridge-gcc default --compiler-gcc default ${CPU_ONLY}
-python ${ENTRY} ${VENV_PATH} -s build_tao_bridge ${CPU_ONLY}
+python ${ENTRY} ${VENV_PATH} -s configure --bridge-gcc default --compiler-gcc default ${CPU_ONLY} --bazel_bridge
+python ${ENTRY} ${VENV_PATH} -s build_tao_bridge ${CPU_ONLY} --bazel_bridge
 python ${ENTRY} ${VENV_PATH} -s build_tao_compiler ${CPU_ONLY}
 python ${ENTRY} ${VENV_PATH} -s build_mlir_ral ${CPU_ONLY}
-python ${ENTRY} ${VENV_PATH} -s test_tao_bridge_cpp ${CPU_ONLY}
-python ${ENTRY} ${VENV_PATH} -s test_tao_bridge_py ${CPU_ONLY}
+python ${ENTRY} ${VENV_PATH} -s test_tao_bridge_cpp ${CPU_ONLY} --bazel_bridge
+python ${ENTRY} ${VENV_PATH} -s test_tao_bridge_py ${CPU_ONLY} --bazel_bridge
 python ${ENTRY} ${VENV_PATH} -s test_tao_compiler ${CPU_ONLY}
 
 # copy libtao_ops.so and tao_compiler_main to blade-disc-tf
-cp tao/build/libtao_ops.so ${BLADE_DISC_DIR}
+cp tao/bazel-bin/libtao_ops.so ${BLADE_DISC_DIR}
 cp tf_community/bazel-bin/tensorflow/compiler/decoupling/tao_compiler_main ${BLADE_DISC_DIR}
 
 (cd tao && \
