+++ tao/.bazelrc
@@ -1,17 +1,13 @@
-# common config
-common --color=yes
-common --experimental_ui_max_stdouterr_bytes=5242880
-common --experimental_repo_remote_exec
-test --jobs=1
-test --test_timeout=100000
-test --verbose_failures
-test --test_output=errors
-build --verbose_failures
-build --spawn_strategy=local  # this is a workaround to include third party source files with BUILD.bazel.
-
 # these comes from tensorflow's source code
+build:release --experimental_repo_remote_exec
+build:release --verbose_failures
+build:release --spawn_strategy=local  # this is a workaround to include third party source files with BUILD.bazel.
 # base
+build:release_base --config=release
 build:release_base --distinct_host_configuration=false
+build:release_base --cxxopt=-std=c++14
+build:release_base --host_cxxopt=-std=c++14
+build:release_base --compilation_mode=opt
 # cpu
 build:avx_linux --copt=-mavx
 build:avx_linux --host_copt=-mavx
@@ -23,22 +19,17 @@ build:release_cpu_linux --crosstool_top="@ubuntu18.04-gcc7_manylinux2010-cuda11.
 build:cuda --repo_env TF_NEED_CUDA=1
 build:cuda --crosstool_top=@local_config_cuda//crosstool:toolchain
 build:cuda --@local_config_cuda//:enable_cuda
+build:cuda --config=release
 # dcu
 build:dcu --crosstool_top=@local_config_rocm//crosstool:toolchain
 build:dcu --define=using_rocm=true --define=using_rocm_hipcc=true --define=using_dcu=true
 build:dcu --repo_env TF_NEED_ROCM=1 --repo_env TF_NEED_DCU=1
-# Config setting to build oneDNN with Compute Library for the Arm Architecture (ACL).
-# This build is for the inference regime only.
-build:mkl_aarch64 --define=build_with_mkl_aarch64=true
-build:mkl_aarch64 --define=tensorflow_mkldnn_contraction_kernel=0
-build:mkl_aarch64 --define=build_with_openmp=true
-build:mkl_aarch64 -c opt
 
 # these comes from tao_compiler, unify these setting in the future
 build:disc --define framework_shared_object=false --experimental_multi_threaded_digest
 build:disc_cpu --config=disc --cxxopt=-DTAO_CPU_ONLY
 build:disc_x86 --config=disc_cpu --config=release_cpu_linux --cxxopt=-DTAO_X86 --define disc_x86=true
-build:disc_aarch64 --config=disc_cpu --config=mkl_aarch64 --cxxopt=-DTAO_AARCH64 --define disc_aarch64=true --linkopt="-Xlinker --stub-group-size -Xlinker 10000000"
+build:disc_aarch64 --config=disc_cpu --cxxopt=-DTAO_AARCH64 --define disc_aarch64=true --linkopt="-Xlinker --stub-group-size -Xlinker 10000000"
 build:disc_cuda --config=disc --config=cuda
 build:disc_dcu --config=disc --config=dcu
 
@@ -47,4 +38,8 @@ build:disc_blaze --config=disc --cxxopt=-DBLAZE_OPT
 
 build:cxx11abi_0 --cxxopt=-D_GLIBCXX_USE_CXX11_ABI=0
 build:cxx11abi_1 --cxxopt=-D_GLIBCXX_USE_CXX11_ABI=1
+
+build:platform_alibaba --cxxopt=-DPLATFORM_ALIBABA
+build:platform_alibaba --define is_platform_alibaba=true
+
 try-import %workspace%/.bazelrc_gen
